added ability to purchase a certificate

+++ repository/src/main/java/com/epam/esm/entity/CertificateTagConnecting.java
@@ -1,11 +1,11 @@
 package com.epam.esm.entity;
 
 public class CertificateTagConnecting {
-    private int id;
-    private int certificateId;
-    private int tagId;
+    private long id;
+    private long certificateId;
+    private long tagId;
 
-    public CertificateTagConnecting(int certificateId, int tagId) {
+    public CertificateTagConnecting(long certificateId, long tagId) {
         this.certificateId = certificateId;
         this.tagId = tagId;
     }
@@ -13,27 +13,27 @@ public class CertificateTagConnecting {
     public CertificateTagConnecting() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getCertificateId() {
+    public long getCertificateId() {
         return certificateId;
     }
 
-    public void setCertificateId(int certificateId) {
+    public void setCertificateId(long certificateId) {
         this.certificateId = certificateId;
     }
 
-    public int getTagId() {
+    public long getTagId() {
         return tagId;
     }
 
-    public void setTagId(int tagId) {
+    public void setTagId(long tagId) {
         this.tagId = tagId;
     }
 }
